recipes: ipsec_esp_ah_comp type conversion
recipes/regression_tests/phase3/ipsec_esp_ah_comp.py
=============================================
division in py3 by default returns float, however,
key is expecting an integer. Casting float to int

+++ recipes/regression_tests/phase3/ipsec_esp_ah_comp.py
@@ -16,7 +16,7 @@ ciphers = []
 #lenth param is in bits
 def generate_key(length):
     key = "0x"
-    key = key + (length/8) * "0b"
+    key = key + int(length/8) * "0b"
     return key
 
 ciphers.append(('aes', 128))
